Seperate autonomous and user control
This will be more useful once we have a longer autonomous
by seperating the files we make it easier to quickly edit the
one we need, especially on a time crunch like competition.
As for main.h, I'm more trying to learn what I can do,
and later why I can do.