Add prototype autonomous
In the current state, this untested autonomous would probably
destroy the bot if ran. There are no sensors currently programmed
in and the directions may also be wrong. This is just to get
the basic idea down so I have something to work with when I
finally get a chance to work with the bot