Strip down regionals code for use on the states bot
The states bot does not have a lot of features yet, so this is
a strip down from the regional bot's code so that it can be run
on the states bot in its current state. More features will be added
later. If anything, the only important file was motion.h for pairings.